refactor: start replacing legacy parser functions with dependent data sources
