+++ src/data/source/implementation/PlaceActivePlayersDataSource.php
@@ -18,36 +18,49 @@
  * @file
  */
 
-namespace MediaWiki\Extension\RobloxAPI\parserFunction;
+namespace MediaWiki\Extension\RobloxAPI\data\source\implementation;
 
+use MediaWiki\Extension\RobloxAPI\data\args\ArgumentSpecification;
 use MediaWiki\Extension\RobloxAPI\data\source\DataSourceProvider;
-use MediaWiki\Extension\RobloxAPI\util\RobloxAPIException;
-use MediaWiki\Extension\RobloxAPI\util\RobloxAPIUtil;
+use MediaWiki\Extension\RobloxAPI\data\source\DependentDataSource;
+use Parser;
 
-/**
- * Gets the amount of active players for a game in a universe.
- */
-class ActivePlayersParserFunction extends RobloxApiParserFunction {
+class PlaceActivePlayersDataSource extends DependentDataSource {
 
+	/**
+	 * @inheritDoc
+	 */
 	public function __construct( DataSourceProvider $dataSourceProvider ) {
-		parent::__construct( $dataSourceProvider );
+		parent::__construct( $dataSourceProvider, 'activePlayers', 'gameData' );
 	}
 
 	/**
 	 * @inheritDoc
 	 */
-	public function exec( $parser, ...$args ): string {
-		[ $universeId, $gameId ] = RobloxAPIUtil::safeDestructure( $args, 2 );
-
-		$source = $this->dataSourceProvider->getDataSourceOrThrow( 'gameData' );
-
-		$gameData = $source->fetch( $universeId, $gameId );
+	public function exec(
+		DataSourceProvider $dataSourceProvider, Parser $parser, array $requiredArgs, array $optionalArgs = []
+	) {
+		$gameData = $this->dataSource->exec( $dataSourceProvider, $parser, $requiredArgs );
 
 		if ( !$gameData ) {
-			throw new RobloxAPIException( 'robloxapi-error-datasource-returned-no-data' );
+			return $this->failNoData();
 		}
 
 		return $gameData->playing;
+	}
+
+	/**
+	 * @inheritDoc
+	 */
+	public function getArgumentSpecification(): ArgumentSpecification {
+		return new ArgumentSpecification( [ 'UniverseID', 'GameID' ] );
+	}
+
+	/**
+	 * @inheritDoc
+	 */
+	public function shouldRegisterLegacyParserFunction(): bool {
+		return true;
 	}
 
 }
