--- src/parserFunction/GroupRankParserFunction.php
@@ -1,46 +1,0 @@
-<?php
-/**
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program; if not, write to the Free Software Foundation, Inc.,
- * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
- * http://www.gnu.org/copyleft/gpl.html
- *
- * @file
- */
-
-namespace MediaWiki\Extension\RobloxAPI\parserFunction;
-
-use MediaWiki\Extension\RobloxAPI\data\source\DataSourceProvider;
-use MediaWiki\Extension\RobloxAPI\util\RobloxAPIUtil;
-
-class GroupRankParserFunction extends RobloxApiParserFunction {
-
-	public function __construct( DataSourceProvider $dataSourceProvider ) {
-		parent::__construct( $dataSourceProvider );
-	}
-
-	/**
-	 * @inheritDoc
-	 */
-	public function exec( $parser, ...$args ): string {
-		[ $groupId, $userId ] = RobloxAPIUtil::safeDestructure( $args, 2 );
-
-		// validate here since the group id is not passed to the data source
-		RobloxAPIUtil::assertValidIds( $groupId, $userId );
-
-		$source = $this->dataSourceProvider->getDataSourceOrThrow( 'groupRank' );
-
-		return $source->exec( $this->dataSourceProvider, $parser, [ $groupId, $userId ] );
-	}
-
-}
